+++ src/Mpociot/BotMan/Drivers/Driver.php
@@ -2,6 +2,7 @@
 
 namespace Mpociot\BotMan\Drivers;
 
+use Mpociot\BotMan\Message;
 use Illuminate\Support\Collection;
 use Mpociot\BotMan\Interfaces\HttpInterface;
 use Symfony\Component\HttpFoundation\Request;
@@ -26,6 +27,15 @@ abstract class Driver implements DriverInterface
         $this->http = $http;
         $this->config = Collection::make($config);
         $this->buildPayload($request);
+    }
+
+    /**
+     * @param Message $matchingMessage
+     * @return void
+     */
+    public function types(Message $matchingMessage)
+    {
+        // Do nothing
     }
 
     /**